# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Food API"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Get all products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "All products"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product",
                "description": "Create a product",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "api_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "brand": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "category": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "tags": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ingredients": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "serving_size_unit": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "serving_size": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "calories": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "fat": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "carbohydrates": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "protein": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by ID",
                "description": "Get a product by ID",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Update a product",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "api_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "brand": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "category": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "tags": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ingredients": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "serving_size_unit": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "serving_size": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "calories": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "fat": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "carbohydrates": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    },
                                    "protein": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Delete a product",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}